+++ crates/paimon/tests/pk_vector_baseline_test.rs
@@ -423,7 +423,7 @@ async fn build_table_with_first_row_id(
     let index_file = IndexFileMeta {
         index_type: INDEX_TYPE.to_string(),
         file_name: index_file_name,
-        file_size: i32::try_from(index_file_size).unwrap(),
+        file_size: i64::try_from(index_file_size).unwrap(),
         row_count: i32::try_from(row_count).unwrap(),
         deletion_vectors_ranges: None,
         global_index_meta: Some(GlobalIndexMeta {
@@ -1270,7 +1270,7 @@ async fn pk_vector_refine_factor_matches_exact_ground_truth() {
     let index_file = IndexFileMeta {
         index_type: INDEX_TYPE.to_string(),
         file_name: index_file_name,
-        file_size: i32::try_from(index_file_size).unwrap(),
+        file_size: i64::try_from(index_file_size).unwrap(),
         row_count: i32::try_from(row_count).unwrap(),
         deletion_vectors_ranges: None,
         global_index_meta: Some(GlobalIndexMeta {
